Fixed a missing cast

+++ CD2_Bot/CD2_Bot/Gear.cs
@@ -234,7 +234,7 @@ namespace CD2_Bot
 
                 case "Berserker's Fury":
                     character.HP = character.MaxHP/2;
-                    weapon.BaseDamage += character.HP/3.34;
+                    weapon.BaseDamage += (int)(character.HP/3.34);
                     break;
 
                 case "Chilly":
